feat: add creating goal functionality

+++ app/src/main/java/com/example/financialtracker/data/Goal.kt
@@ -2,13 +2,13 @@ package com.example.financialtracker.data
 
 data class Goal(
     val name: String,
-    val currentAmount: Int,
-    val targetAmount: Int,
+    val currentAmount: Double,
+    val targetAmount: Double,
     val deadline: String
 )
 
 data class CreateGoal(
     val name: String,
-    val targetAmount: Int,
+    val targetAmount: Double,
     val deadline: String
 )
